requirements.txt
absl-py==2.3.1
agno==2.5.17
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
audioop-lts==0.2.2
av==16.1.0
babel==2.17.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.2.0
blinker==1.9.0
blis==1.3.3
build==1.4.3
cachetools==6.1.0
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.7
click==8.2.1
cloudpathlib==0.23.0
cloup==3.0.8
colorlog==6.10.1
comm==0.2.2
confection==1.3.2
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
cymem==2.0.13
Cython==3.2.4
dataclasses-json==0.6.7
datasets==4.5.0
ddgs==9.13.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
durationpy==0.10
emoji==2.15.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.58.5
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.45
glcontext==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.191.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.66.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
groq==0.37.1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hf-xet==1.4.3
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.10.1
hyperframe==6.1.0
idna==3.10
imbalanced-learn==0.14.0
imblearn==0.0
importlib_metadata==8.7.1
importlib_resources==7.1.0
installer==0.7.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
isosurfaces==0.1.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==3.13.1
keras-tuner==1.4.8
kiwisolver==1.4.8
kt-legacy==1.0.5
kubernetes==35.0.0
langchain==1.2.15
langchain-chroma==1.1.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.28
langchain-groq==1.1.2
langchain-huggingface==1.2.1
langchain-ollama==1.1.0
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langchain-unstructured==1.0.1
langdetect==1.0.9
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.30
libclang==18.1.1
llvmlite==0.47.0
lxml==6.0.4
Mako==1.3.10
manim==0.20.1
ManimPango==0.6.1
mapbox_earcut==2.0.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.4
mmh3==5.2.1
moderngl==5.12.0
moderngl-window==3.1.1
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.15
mypy_extensions==1.1.0
mysql-connector-python==9.4.0
namex==0.1.0
narwhals==2.1.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
notebook==7.4.3
notebook_shim==0.2.4
numba==0.65.0
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
ollama==0.6.1
onnxruntime==1.24.4
openai==2.31.0
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
opt_einsum==3.4.0
optree==0.18.0
optuna==4.6.0
orjson==3.11.8
ormsgpack==1.12.2
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.2
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
preshed==3.0.13
primp==1.2.2
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycairo==1.29.0
pycparser==2.22
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
pydeck==0.9.1
pydub==0.25.1
pyglet==2.1.13
pyglm==2.8.3
Pygments==2.19.2
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyparsing==3.2.3
pypdf==6.10.0
pypdfium2==5.7.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-iso639==2026.1.31
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.26
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
RapidFuzz==3.14.5
referencing==0.36.2
regex==2026.2.28
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.2.0
rpds-py==0.25.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.2
screeninfo==0.8.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==5.4.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
skia-pathops==0.9.2
smart_open==7.5.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.13
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.45
srsly==2.5.3
srt==3.5.3
stack-data==0.6.3
starlette==0.47.2
statsmodels==0.14.6
streamlit==1.48.1
svgelements==1.9.6
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
terminado==0.18.1
thinc==8.3.13
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.1
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
typer==0.24.1
typer-slim==0.21.1
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.22.18
unstructured-client==0.43.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uuid_utils==0.14.1
uv==0.11.6
uvicorn==0.35.0
uvloop==0.22.1
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
weasel==1.0.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==16.0
Werkzeug==3.1.5
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==2.1.2
xgboost==3.1.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
